Move shapes folder to annotator folder
