+++ ui/src/features/annotator/shapes/interfaces.ts
