+++ ui/src/features/annotator/shapes/polygon.component.tsx
