+++ ui/src/features/annotator/shapes/rectangle.component.tsx
